move map replated stuff to mapglhelpers
